Move files from model to domain/entities
